+++ test/mocha/001-service.js
@@ -5,8 +5,10 @@
 
 const async = require('async');
 const bedrock = require('bedrock');
+const brTest = require('bedrock-test');
 const config = bedrock.config;
 const helpers = require('./helpers');
+const jsonpatch = require('fast-json-patch');
 const mockData = require('./mock.data');
 let request = require('request');
 const uuid = require('uuid/v4');
@@ -17,7 +19,7 @@ const identityService = config.server.baseUri +
 
 describe('bedrock-identity-http', function() {
   before(function(done) {
-    helpers.prepareDatabase({mockData: mockData}, done);
+    helpers.prepareDatabase({mockData}, done);
   });
   describe('identity-service', function() {
     describe('unauthenticated requests', function() {
@@ -27,8 +29,8 @@ describe('bedrock-identity-http', function() {
             url: mockData.identities.registered.identity.id,
             method: 'GET'
           },
-          function(err, res, body) {
-            should.not.exist(err);
+          (err, res, body) => {
+            assertNoError(err);
             res.statusCode.should.equal(200);
             should.exist(body);
             body.should.be.an('object');
@@ -48,7 +50,7 @@ describe('bedrock-identity-http', function() {
             method: 'GET'
           },
           function(err, res, body) {
-            should.not.exist(err);
+            assertNoError(err);
             res.statusCode.should.equal(404);
             body.should.be.an('object');
             should.exist(body.type);
@@ -66,15 +68,12 @@ describe('bedrock-identity-http', function() {
 
         describe('GET request for all identities', function() {
           it('should include all identity data', function(done) {
-            const privateKeyPem = actor.keys.privateKey.privateKeyPem;
-            const publicKeyId = actor.keys.privateKey.publicKey;
             sendRequest({
               url: identityService,
               method: 'GET',
-              privateKeyPem: privateKeyPem,
-              publicKeyId: publicKeyId
-            }, function(err, res, body) {
-              should.not.exist(err);
+              authIdentity: actor,
+            }, (err, res, body) => {
+              assertNoError(err);
               res.statusCode.should.equal(200);
               should.exist(body);
               body.should.be.an('array');
@@ -93,25 +92,21 @@ describe('bedrock-identity-http', function() {
               r.url.should.equal(
                 mockData.identities.registered.identity.url);
               r.type.should.equal('Identity');
-              should.exist(r.sysResourceRole);
-              done(err);
+              done();
             });
           });
         });
         describe('POST to create new identity', function() {
           it('should reject a malformed identity');
-          it('should create and lookup a new identity', function(done) {
+          it('should create and lookup a new identity', done => {
             const userId = uuid();
-            const privateKeyPem = actor.keys.privateKey.privateKeyPem;
-            const publicKeyId = actor.keys.privateKey.publicKey;
             async.waterfall([
               function(callback) {
                 sendRequest({
+                  authIdentity: actor,
                   url: identityService,
                   method: 'POST',
                   identity: createIdentity(userId),
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
                 }, callback);
               },
               function(res, body, callback) {
@@ -135,7 +130,7 @@ describe('bedrock-identity-http', function() {
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
@@ -163,11 +158,11 @@ describe('bedrock-identity-http', function() {
                 should.exist(body);
                 body.should.be.an('object');
                 should.exist(body.type);
-                body.type.should.equal('AddIdentityFailed');
+                body.type.should.equal('NotAllowedError');
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
@@ -175,28 +170,24 @@ describe('bedrock-identity-http', function() {
           it('should return 409 on a duplicate identity', function(done) {
             const userId = uuid();
             const actor = mockData.identities.adminUser;
-            const privateKeyPem = actor.keys.privateKey.privateKeyPem;
-            const publicKeyId = actor.keys.privateKey.publicKey;
             async.waterfall([
               function(callback) {
                 sendRequest(
                   {
+                    authIdentity: actor,
                     url: identityService,
                     method: 'POST',
                     identity: createIdentity(userId),
-                    privateKeyPem: privateKeyPem,
-                    publicKeyId: publicKeyId
                   }, callback);
               },
               function(res, body, callback) {
                 res.statusCode.should.equal(201);
                 sendRequest(
                   {
+                    authIdentity: actor,
                     url: identityService,
                     method: 'POST',
                     identity: createIdentity(userId),
-                    privateKeyPem: privateKeyPem,
-                    publicKeyId: publicKeyId
                   }, callback);
               },
               function(res, body, callback) {
@@ -205,7 +196,7 @@ describe('bedrock-identity-http', function() {
                 body.should.be.an('object');
                 should.exist(body.type);
                 body.type.should.be.a('string');
-                body.type.should.equal('DuplicateIdentity');
+                body.type.should.equal('DuplicateError');
                 should.exist(body.details);
                 body.details.should.be.an('object');
                 should.exist(body.details.identity);
@@ -213,7 +204,7 @@ describe('bedrock-identity-http', function() {
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
@@ -223,8 +214,6 @@ describe('bedrock-identity-http', function() {
           const actor = mockData.identities.adminUser;
           const userId = uuid();
           const groupId = uuid();
-          const privateKeyPem = actor.keys.privateKey.privateKeyPem;
-          const publicKeyId = actor.keys.privateKey.publicKey;
           let createdIdentity = null;
           let createdGroup = null;
 
@@ -235,29 +224,27 @@ describe('bedrock-identity-http', function() {
               generateResource: 'id'
             }];
             async.auto({
-              createIdentity: function(callback) {
+              createIdentity: callback => {
                 sendRequest({
+                  authIdentity: actor,
                   url: identityService,
                   method: 'POST',
                   identity: newIdentity,
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
                 }, function(err, res, body) {
                   res.statusCode.should.equal(201);
                   createdIdentity = body;
                   callback(err, createdIdentity);
                 });
               },
-              createGroup: ['createIdentity', function(callback) {
+              createGroup: ['createIdentity', (results, callback) => {
                 const group = createIdentity(groupId);
                 group.owner = actor.id;
                 group.type = ['Identity', 'Group'];
                 sendRequest({
+                  authIdentity: actor,
                   url: identityService,
                   method: 'POST',
                   identity: group,
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
                 }, function(err, res, body) {
                   createdGroup = body;
                   callback(err);
@@ -268,17 +255,21 @@ describe('bedrock-identity-http', function() {
 
           it('should update an identities label', function(done) {
             const newData = uuid();
+            const updatedIdentity = createdIdentity;
+            const observer = jsonpatch.observe(updatedIdentity);
+            updatedIdentity.label = newData;
+            const patch = jsonpatch.generate(observer);
+            jsonpatch.unobserve(updatedIdentity, observer);
             async.waterfall([
               function(callback) {
                 sendRequest({
+                  authIdentity: actor,
                   url: createdIdentity.id,
                   method: 'PATCH',
-                  identity: [{
-                    changes: {label: newData},
-                    op: 'updateIdentity'
-                  }],
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
+                  identity: {
+                    patch,
+                    sequence: 0
+                  },
                 }, callback);
               },
               function(res, body, callback) {
@@ -298,24 +289,28 @@ describe('bedrock-identity-http', function() {
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
 
-          it('should update an identities description', function(done) {
+          it('should update an identities description', done => {
             const newData = uuid();
+            const updatedIdentity = createdIdentity;
+            const observer = jsonpatch.observe(updatedIdentity);
+            updatedIdentity.description = newData;
+            const patch = jsonpatch.generate(observer);
+            jsonpatch.unobserve(updatedIdentity, observer);
             async.waterfall([
               function(callback) {
                 sendRequest({
+                  authIdentity: actor,
                   url: createdIdentity.id,
                   method: 'PATCH',
-                  identity: [{
-                    changes: {description: newData},
-                    op: 'updateIdentity'
-                  }],
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
+                  identity: {
+                    patch,
+                    sequence: 1
+                  },
                 }, callback);
               },
               function(res, body, callback) {
@@ -335,12 +330,13 @@ describe('bedrock-identity-http', function() {
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
 
-          it('should update an identities sysResourceRoles', function(done) {
+          // TODO: implement updateRoles API
+          it.skip('should update an identities sysResourceRoles', done => {
             const updatedSysResourceRole = actor.identity.sysResourceRole;
             updatedSysResourceRole.push({
               sysRole: 'bedrock-identity-http.identity.manager',
@@ -349,6 +345,7 @@ describe('bedrock-identity-http', function() {
             async.waterfall([
               function(callback) {
                 sendRequest({
+                  authIdentity: actor,
                   url: actor.identity.id,
                   method: 'PATCH',
                   identity: [{
@@ -357,18 +354,15 @@ describe('bedrock-identity-http', function() {
                     },
                     op: 'add'
                   }],
-                  privateKeyPem: privateKeyPem,
-                  publicKeyId: publicKeyId
                 }, callback);
               },
               function(res, body, callback) {
                 res.statusCode.should.equal(204);
                 sendRequest(
                   {
+                    authIdentity: actor,
                     url: actor.identity.id,
                     method: 'GET',
-                    privateKeyPem: privateKeyPem,
-                    publicKeyId: publicKeyId
                   }, callback);
               },
               function(res, body, callback) {
@@ -379,7 +373,7 @@ describe('bedrock-identity-http', function() {
                 callback();
               }
             ], function(err) {
-              should.not.exist(err);
+              assertNoError(err);
               done();
             });
           });
@@ -397,26 +391,36 @@ describe('bedrock-identity-http', function() {
 function sendRequest(options, callback) {
   const url = options.url;
   const reqBody = options.identity || null;
-  const privateKeyPem = options.privateKeyPem || null;
-  const publicKeyId = options.publicKeyId || null;
   const reqMethod = options.method;
+  const authIdentity = options.authIdentity || null;
 
   const requestOptions = {
     method: reqMethod,
     url: url,
     json: true
   };
-  if(privateKeyPem) {
-    requestOptions.httpSignature = {
-      key: privateKeyPem,
-      keyId: publicKeyId,
-      headers: ['date', 'host', 'request-line']
-    };
-  }
+
   if(reqBody) {
     requestOptions.body = reqBody;
   }
-  request(requestOptions, callback);
+  async.auto({
+    sign: callback => {
+      if(!authIdentity) {
+        return callback();
+      }
+      // mutates requestOptions
+      brTest.helpers.createHttpSignatureRequest(
+        {algorithm: 'rsa-sha256', identity: authIdentity, requestOptions},
+        callback);
+    },
+    request: ['sign', (results, callback) => request(requestOptions, callback)]
+  }, (err, results) => {
+    if(err) {
+      return callback(err);
+    }
+    const [res, body] = results.request;
+    callback(err, res, body);
+  });
 }
 
 function createIdentity(userName) {