+++ test/test.config.js
@@ -26,10 +26,9 @@ roles['bedrock-identity-http.identity.administrator'] = {
   label: 'Identity Administrator',
   comment: 'Role for identity administrators.',
   sysPermission: [
-    permissions.IDENTITY_ADMIN.id,
     permissions.IDENTITY_ACCESS.id,
     permissions.IDENTITY_INSERT.id,
-    permissions.IDENTITY_EDIT.id,
+    permissions.IDENTITY_UPDATE.id,
     permissions.IDENTITY_REMOVE.id,
     permissions.IDENTITY_CAPABILITY_DELEGATE.id,
     permissions.PUBLIC_KEY_CREATE.id,
@@ -43,7 +42,7 @@ roles['bedrock-identity-http.identity.manager'] = {
   sysPermission: [
     permissions.IDENTITY_ACCESS.id,
     permissions.IDENTITY_INSERT.id,
-    permissions.IDENTITY_EDIT.id,
+    permissions.IDENTITY_UPDATE.id,
     permissions.IDENTITY_CAPABILITY_DELEGATE.id,
     permissions.PUBLIC_KEY_CREATE.id,
     permissions.PUBLIC_KEY_REMOVE.id